+++ examples/next-agent/agent/weather-agent.ts
@@ -1,10 +1,6 @@
 import { weatherTool } from '@/tool/weather-tool';
 import { openai } from '@ai-sdk/openai';
-import {
-  Experimental_Agent as Agent,
-  Experimental_InferAgentUIMessage as InferAgentUIMessage,
-  stepCountIs,
-} from 'ai';
+import { Agent, InferAgentUIMessage } from 'ai';
 
 export const weatherAgent = new Agent({
   model: openai('gpt-4o'),
@@ -12,7 +8,6 @@ export const weatherAgent = new Agent({
   tools: {
     weather: weatherTool,
   },
-  stopWhen: stepCountIs(10),
 });
 
 export type WeatherAgentUIMessage = InferAgentUIMessage<typeof weatherAgent>;