+++ packages/ai/src/types/language-model.ts
@@ -1,4 +1,5 @@
 import {
+  LanguageModelV2,
   LanguageModelV3,
   LanguageModelV3CallWarning,
   LanguageModelV3FinishReason,
@@ -8,7 +9,7 @@ import {
 /**
 Language model that is used by the AI SDK Core functions.
 */
-export type LanguageModel = string | LanguageModelV3;
+export type LanguageModel = string | LanguageModelV3 | LanguageModelV2;
 
 /**
 Reason why a language model finished generating a response.